+++ recovery/PreGANSrc/src/train_multiobjective.py
@@ -88,26 +88,26 @@ def train_gan_multiobjective(gen, disc, gopt, dopt, embedding, schedule_data, en
     new_score = 0.8 * new_energy + 0.2 * new_response_time + 0.01 * actual_migration_count
     orig_score = 0.8 * orig_energy + 0.2 * orig_response_time + 0.01 * calculate_migration_count(schedule_data, schedule_data)
     true_class = torch.tensor([0, 1] if new_score <= orig_score else [1, 0], 
-                              dtype=torch.double, device=class_probs.device)
+                              dtype=torch.float32, device=class_probs.device)
     class_loss = ganloss(class_probs, true_class)
     
     # Task 2: Energy prediction loss
-    energy_target = torch.tensor([new_energy], dtype=torch.double, device=energy_pred.device)
+    energy_target = torch.tensor([new_energy], dtype=torch.float32, device=energy_pred.device)
     energy_loss = mse_loss(energy_pred, energy_target)
     
     # Task 3: Response time prediction loss
-    response_time_target = torch.tensor([new_response_time], dtype=torch.double, device=response_time_pred.device)
+    response_time_target = torch.tensor([new_response_time], dtype=torch.float32, device=response_time_pred.device)
     response_time_pred_loss = mse_loss(response_time_pred, response_time_target)
     
     # Task 4: Migration cost prediction loss
-    migration_cost_target = torch.tensor([actual_migration_count], dtype=torch.double, device=migration_cost_pred.device)
+    migration_cost_target = torch.tensor([actual_migration_count], dtype=torch.float32, device=migration_cost_pred.device)
     migration_cost_pred_loss = mse_loss(migration_cost_pred, migration_cost_target)
 
     # Normalize regression losses to comparable scales to avoid DLoss explosion
     # Typical scales: energy ~ 2e4 (MSE ~ O(1e8)), rt ~ O(1e3-1e4) w.r.t SLA (MSE ~ large), mc ~ O(1e2)
-    energy_norm_denom = (torch.tensor([max(new_energy, orig_energy)], dtype=torch.double, device=energy_pred.device) ** 2)
-    rt_norm_denom = (torch.tensor([sla_threshold], dtype=torch.double, device=response_time_pred.device) ** 2)
-    mc_norm_denom = (torch.tensor([migration_cost_threshold], dtype=torch.double, device=migration_cost_pred.device) ** 2)
+    energy_norm_denom = (torch.tensor([max(new_energy, orig_energy)], dtype=torch.float32, device=energy_pred.device) ** 2)
+    rt_norm_denom = (torch.tensor([sla_threshold], dtype=torch.float32, device=response_time_pred.device) ** 2)
+    mc_norm_denom = (torch.tensor([migration_cost_threshold], dtype=torch.float32, device=migration_cost_pred.device) ** 2)
     energy_loss_norm = energy_loss / energy_norm_denom
     response_time_pred_loss_norm = response_time_pred_loss / rt_norm_denom
     migration_cost_pred_loss_norm = migration_cost_pred_loss / mc_norm_denom
@@ -130,36 +130,36 @@ def train_gan_multiobjective(gen, disc, gopt, dopt, embedding, schedule_data, en
     
     # Generator loss: balance multiple objectives
     # Method 1: Classification loss
-    target_better = torch.tensor([0, 1], dtype=torch.double, device=class_probs_gen.device)
+    target_better = torch.tensor([0, 1], dtype=torch.float32, device=class_probs_gen.device)
     gen_class_loss = ganloss(class_probs_gen, target_better)
     
     # Method 2: Energy constraint loss (encourage predicting lower energy)
-    energy_upper_bound = torch.tensor([orig_energy], dtype=torch.double, device=energy_pred_gen.device)
+    energy_upper_bound = torch.tensor([orig_energy], dtype=torch.float32, device=energy_pred_gen.device)
     # Normalize energy constraint by upper bound to keep scale reasonable
     gen_energy_loss = torch.relu(energy_pred_gen - energy_upper_bound + 0.1) / energy_upper_bound
     
     # Method 3: Response time constraint loss (penalize exceeding SLA threshold)
-    sla_threshold_tensor = torch.tensor([sla_threshold], dtype=torch.double, device=response_time_pred_gen.device)
+    sla_threshold_tensor = torch.tensor([sla_threshold], dtype=torch.float32, device=response_time_pred_gen.device)
     response_time_excess = torch.relu(response_time_pred_gen - sla_threshold_tensor)
     # Normalize RT constraint by SLA so its magnitude is comparable
     gen_response_time_loss = response_time_weight * (response_time_excess / sla_threshold_tensor)
     
     # Also penalize actual response time
-    actual_response_time_tensor = torch.tensor([new_response_time], dtype=torch.double, device=response_time_pred_gen.device)
+    actual_response_time_tensor = torch.tensor([new_response_time], dtype=torch.float32, device=response_time_pred_gen.device)
     actual_response_time_excess = torch.relu(actual_response_time_tensor - sla_threshold_tensor)
     gen_actual_response_time_loss = response_time_weight * 0.5 * (actual_response_time_excess / sla_threshold_tensor)
     
     # Method 4: Migration cost constraint loss (进一步增强版，关键)
     # Penalize if predicted migration cost exceeds threshold
     migration_cost_threshold_tensor = torch.tensor([migration_cost_threshold], 
-                                                    dtype=torch.double, device=migration_cost_pred_gen.device)
+                                                    dtype=torch.float32, device=migration_cost_pred_gen.device)
     migration_cost_excess = torch.relu(migration_cost_pred_gen - migration_cost_threshold_tensor)
     # 进一步增强迁移成本约束：使用立方惩罚，使超过阈值时惩罚更严重
     gen_migration_cost_loss = migration_cost_weight * (migration_cost_excess ** 3 + migration_cost_excess ** 2 + migration_cost_excess)
     
     # Also penalize actual migration cost (进一步增强版)
     actual_migration_cost_tensor = torch.tensor([actual_migration_count], 
-                                                dtype=torch.double, device=migration_cost_pred_gen.device)
+                                                dtype=torch.float32, device=migration_cost_pred_gen.device)
     actual_migration_cost_excess = torch.relu(actual_migration_cost_tensor - migration_cost_threshold_tensor)
     # Phase 1 optimization: 进一步增强实际迁移成本约束：使用立方惩罚，权重从1.5增加到2.0（更严格）
     gen_actual_migration_cost_loss = migration_cost_weight * 2.0 * (actual_migration_cost_excess ** 3 + actual_migration_cost_excess ** 2 + actual_migration_cost_excess)
